# Fix performance-noexcept-move-constructor clang-tidy warning