--- a/.clang-tidy
+++ b/.clang-tidy
@@ -58,7 +58,6 @@ Checks: >
     modernize-concat-nested-namespaces,
     modernize-use-using,
     performance-*,
-    -performance-noexcept-move-constructor,
     -performance-no-int-to-ptr,
     -performance-enum-size,
     -performance-avoid-endl,