<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1102.58<br/>plain u32: 1094.19<br/>aliased double: 953.45<br/>plain double: 950.449 | aliased u32: 1090<br/>plain u32: 1092.31<br/>aliased double: 959.847<br/>plain double: 952.227 |
| e2e_match_ch | requests: 344<br/>failures: 0<br/>req/s: 5.734req/s<br/>avg: 5.349ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 15ms<br/>98%: 48ms<br/>99%: 80ms   <br/>min: 1.274ms<br/>max: 95.968ms | requests: 361<br/>failures: 0<br/>req/s: 6.021req/s<br/>avg: 4.955ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 15ms<br/>98%: 26ms<br/>99%: 52ms   <br/>min: 1.249ms<br/>max: 101.692ms |
| e2e_match_mld | requests: 359<br/>failures: 0<br/>req/s: 5.995req/s<br/>avg: 4.669ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 13ms<br/>98%: 19ms<br/>99%: 37ms   <br/>min: 1.495ms<br/>max: 53.106ms | requests: 371<br/>failures: 0<br/>req/s: 6.186req/s<br/>avg: 4.972ms<br/>50%: 3ms<br/>75%: 4ms<br/>95%: 17ms<br/>98%: 33ms<br/>99%: 46ms   <br/>min: 1.258ms<br/>max: 66.049ms |
| e2e_nearest_ch | requests: 422<br/>failures: 0<br/>req/s: 7.034req/s<br/>avg: 1.553ms<br/>50%: 2ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 2ms<br/>99%: 2ms   <br/>min: 1.007ms<br/>max: 3.340ms | requests: 444<br/>failures: 0<br/>req/s: 7.405req/s<br/>avg: 1.521ms<br/>50%: 1ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 2ms<br/>99%: 2ms   <br/>min: 1.072ms<br/>max: 4.366ms |
| e2e_nearest_mld | requests: 422<br/>failures: 0<br/>req/s: 7.047req/s<br/>avg: 1.696ms<br/>50%: 2ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 3ms<br/>99%: 3ms   <br/>min: 1.140ms<br/>max: 4.738ms | requests: 422<br/>failures: 0<br/>req/s: 7.037req/s<br/>avg: 1.523ms<br/>50%: 1ms<br/>75%: 2ms<br/>95%: 2ms<br/>98%: 2ms<br/>99%: 2ms   <br/>min: 1.076ms<br/>max: 3.124ms |
| e2e_route_ch | requests: 396<br/>failures: 0<br/>req/s: 6.601req/s<br/>avg: 4.409ms<br/>50%: 4ms<br/>75%: 5ms<br/>95%: 6ms<br/>98%: 7ms<br/>99%: 10ms   <br/>min: 1.242ms<br/>max: 135.372ms | requests: 389<br/>failures: 0<br/>req/s: 6.488req/s<br/>avg: 4.061ms<br/>50%: 4ms<br/>75%: 5ms<br/>95%: 6ms<br/>98%: 7ms<br/>99%: 9ms   <br/>min: 1.495ms<br/>max: 10.259ms |
| e2e_route_mld | requests: 382<br/>failures: 0<br/>req/s: 6.379req/s<br/>avg: 5.592ms<br/>50%: 5ms<br/>75%: 5ms<br/>95%: 7ms<br/>98%: 10ms<br/>99%: 38ms   <br/>min: 1.738ms<br/>max: 133.708ms | requests: 404<br/>failures: 0<br/>req/s: 6.737req/s<br/>avg: 4.228ms<br/>50%: 4ms<br/>75%: 5ms<br/>95%: 6ms<br/>98%: 7ms<br/>99%: 8ms   <br/>min: 1.391ms<br/>max: 10.586ms |
| e2e_table_ch | requests: 380<br/>failures: 0<br/>req/s: 6.334req/s<br/>avg: 45.760ms<br/>50%: 28ms<br/>75%: 76ms<br/>95%: 130ms<br/>98%: 140ms<br/>99%: 140ms   <br/>min: 2.381ms<br/>max: 176.579ms | requests: 383<br/>failures: 0<br/>req/s: 6.388req/s<br/>avg: 40.429ms<br/>50%: 27ms<br/>75%: 55ms<br/>95%: 120ms<br/>98%: 130ms<br/>99%: 130ms   <br/>min: 3.061ms<br/>max: 165.447ms |
| e2e_table_mld | requests: 378<br/>failures: 0<br/>req/s: 6.312req/s<br/>avg: 71.565ms<br/>50%: 73ms<br/>75%: 100ms<br/>95%: 130ms<br/>98%: 140ms<br/>99%: 150ms   <br/>min: 4.543ms<br/>max: 168.404ms | requests: 380<br/>failures: 0<br/>req/s: 6.337req/s<br/>avg: 53.289ms<br/>50%: 38ms<br/>75%: 83ms<br/>95%: 130ms<br/>98%: 130ms<br/>99%: 140ms   <br/>min: 2.473ms<br/>max: 154.113ms |
| e2e_trip_ch | requests: 370<br/>failures: 0<br/>req/s: 6.167req/s<br/>avg: 16.001ms<br/>50%: 15ms<br/>75%: 21ms<br/>95%: 30ms<br/>98%: 33ms<br/>99%: 34ms   <br/>min: 4.064ms<br/>max: 36.300ms | requests: 356<br/>failures: 0<br/>req/s: 5.938req/s<br/>avg: 16.339ms<br/>50%: 15ms<br/>75%: 20ms<br/>95%: 31ms<br/>98%: 33ms<br/>99%: 35ms   <br/>min: 4.065ms<br/>max: 44.516ms |
| e2e_trip_mld | requests: 359<br/>failures: 0<br/>req/s: 5.995req/s<br/>avg: 22.065ms<br/>50%: 21ms<br/>75%: 28ms<br/>95%: 35ms<br/>98%: 39ms<br/>99%: 42ms   <br/>min: 5.952ms<br/>max: 84.670ms | requests: 345<br/>failures: 0<br/>req/s: 5.753req/s<br/>avg: 18.188ms<br/>50%: 17ms<br/>75%: 23ms<br/>95%: 32ms<br/>98%: 37ms<br/>99%: 39ms   <br/>min: 3.983ms<br/>max: 40.179ms |
| json-render | String: 6.55045ms<br/>Stringstream: 9.28397ms<br/>Vector: 6.99187ms | String: 6.61233ms<br/>Stringstream: 9.37551ms<br/>Vector: 6.94386ms |
| match_ch | Default radius: <br/>4.40195ms/req at 82 coordinate<br/>0.0536823ms/coordinate<br/>Radius 5m: <br/>4.37023ms/req at 82 coordinate<br/>0.0532955ms/coordinate<br/>Radius 10m: <br/>14.9272ms/req at 82 coordinate<br/>0.182039ms/coordinate<br/>Radius 15m: <br/>36.5714ms/req at 82 coordinate<br/>0.445992ms/coordinate<br/>Radius 30m: <br/>311.882ms/req at 82 coordinate<br/>3.80344ms/coordinate | Default radius: <br/>4.41436ms/req at 82 coordinate<br/>0.0538336ms/coordinate<br/>Radius 5m: <br/>4.39198ms/req at 82 coordinate<br/>0.0535607ms/coordinate<br/>Radius 10m: <br/>15.035ms/req at 82 coordinate<br/>0.183353ms/coordinate<br/>Radius 15m: <br/>36.5207ms/req at 82 coordinate<br/>0.445375ms/coordinate<br/>Radius 30m: <br/>311.698ms/req at 82 coordinate<br/>3.8012ms/coordinate |
| match_mld | Default radius: <br/>2.78667ms/req at 82 coordinate<br/>0.0339838ms/coordinate<br/>Radius 5m: <br/>2.84428ms/req at 82 coordinate<br/>0.0346864ms/coordinate<br/>Radius 10m: <br/>10.3432ms/req at 82 coordinate<br/>0.126136ms/coordinate<br/>Radius 15m: <br/>25.7298ms/req at 82 coordinate<br/>0.313777ms/coordinate<br/>Radius 30m: <br/>302.289ms/req at 82 coordinate<br/>3.68645ms/coordinate | Default radius: <br/>2.75586ms/req at 82 coordinate<br/>0.033608ms/coordinate<br/>Radius 5m: <br/>2.74494ms/req at 82 coordinate<br/>0.0334748ms/coordinate<br/>Radius 10m: <br/>10.1037ms/req at 82 coordinate<br/>0.123216ms/coordinate<br/>Radius 15m: <br/>25.7695ms/req at 82 coordinate<br/>0.314263ms/coordinate<br/>Radius 30m: <br/>300.826ms/req at 82 coordinate<br/>3.6686ms/coordinate |
| osrm_contract | Time: 94.21s Peak RAM: 185.54MB | Time: 93.07s Peak RAM: 185.49MB |
| osrm_customize | Time: 1.31s Peak RAM: 115.11MB | Time: 1.31s Peak RAM: 115.10MB |
| osrm_extract | Time: 12.10s Peak RAM: 404.58MB | Time: 12.10s Peak RAM: 395.95MB |
| osrm_partition | Time: 2.21s Peak RAM: 148.46MB | Time: 2.18s Peak RAM: 148.46MB |
| packedvector | random write:<br/>std::vector 11107.2 ms<br/>util::packed_vector 74876.7 ms<br/>slowdown: 6.7413<br/>random read:<br/>std::vector 11129 ms<br/>util::packed_vector 30968.8 ms<br/>slowdown: 2.78272 | random write:<br/>std::vector 9751.37 ms<br/>util::packed_vector 73639.6 ms<br/>slowdown: 7.55172<br/>random read:<br/>std::vector 8424.02 ms<br/>util::packed_vector 30253.6 ms<br/>slowdown: 3.59135 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>512.899ms<br/>0.512899ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>349.785ms<br/>0.349785ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>625.685ms<br/>0.625685ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>150.064ms<br/>0.150064ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.1439ms<br/>0.0971439ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.73ms<br/>0.13273ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>149.48ms<br/>0.14948ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.2409ms<br/>0.0972409ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>132.173ms<br/>0.132173ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>510.857ms<br/>0.510857ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>350.776ms<br/>0.350776ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>624.471ms<br/>0.624471ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>150.867ms<br/>0.150867ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.0311ms<br/>0.0970311ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>131.516ms<br/>0.131516ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>150.489ms<br/>0.150489ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>96.9338ms<br/>0.0969338ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>131.62ms<br/>0.13162ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>638.024ms<br/>0.638024ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>437.046ms<br/>0.437046ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>810.722ms<br/>0.810722ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>268.606ms<br/>0.268606ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>162.358ms<br/>0.162358ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>292.753ms<br/>0.292753ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>268.794ms<br/>0.268794ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>162.134ms<br/>0.162134ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>293.145ms<br/>0.293145ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>646.06ms<br/>0.64606ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>436.853ms<br/>0.436853ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>812.85ms<br/>0.81285ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>268.441ms<br/>0.268441ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>162.694ms<br/>0.162694ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>287.822ms<br/>0.287822ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>265.668ms<br/>0.265668ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>161.826ms<br/>0.161826ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>287.595ms<br/>0.287595ms/req |
| rtree | 1 result:<br/>207.791ms ->  0.0207791 ms/query<br/>10 results:<br/>242.731ms ->  0.0242731 ms/query | 1 result:<br/>206.85ms ->  0.020685 ms/query<br/>10 results:<br/>241.899ms ->  0.0241899 ms/query |
</details>
<!-- BENCHMARK_RESULTS_END -->